Use http client with keepalive & let user override
Because Dynamo queries are often a very high frequency of fast requests,
the underlying TCP connections should be reused. This uses a default
HttpClient with idle connections kept around for 30 seconds.

It also adds a field on the Server struct to let a user pass in their
own client configured however they'd like.